Quickfix `_count` endpoint generated SQL (#1261)
* Fixes https://github.com/QuesmaOrg/quesma/issues/985

I use `fmt.Sprintf("...")` for creating the full query, instead of
`queryRow("select ... from ?", tableName)`. It's a bit worse, as passing
tablename as an argument to the driver should be preferred, but driver
escapes every string argument with \` It's bad, because ``FROM `table`
`` will work, but , but ``FROM `db.table` `` does not. To not waste more
time toying with the driver, etc, I propose to add such a quickfix.

* Adds a simple test for `_count` endpoint - if I'm not mistaken, we had
none before.