Turing-2 Lucene parser should return `iLIKE '%abc%'`, not `= 'abc'` (#1228)
We generate `log = 'abc'`, but it should be `iLIKE '%abc%'`, as from
doc:

> where the author field CONTAINS the exact phrase "john smith"
> author:"John Smith"

The change is small, but it wasn't that straightforward to get it all
right, especially with some special characters.
I did a few manual "hard" tests with Kibana/Clickhouse, and everything
seemed to work fine, but because of the above, maybe some polishing will
need to be done in the future.

Also added a few new tests for some trickier inputs.