Add tests of lexer dialects (sqlparse, sqlfluff)
This PR adds tests of our ports of dialects from sqlparse and sqlfluff.

The test files are the extracted SQL queries from the tests of sqlparse
and sqlfluff. The expected outputs (tokens) are collected by running
the queries through the original lexers of sqlparse and sqlfluff.

By doing this, we get a really good test coverage (>10k SQL queries,
many edge case/"strange"/"tricky" SQL queries collected) and we can
be almost certain that our implementation behaves identically to the
original sqlparse/sqlfluff implementations.

The test files are stored in a separate repo (added to Quesma as a Git
submodule), because they are large (~10MB).