[`sample_flights`] Fix `total` in `top_hits` response (#1337)
Makes the last skipped `sample_flights` test pass.
In other aggregations we have `parent_count` for that. Here I don't add
a new column to select, just extract from the existing ones, as it's
always there and I find it much simpler to fix it this way. We can
change that later if needed.