[minor QOL] Unify testname throughout unit tests (#1379)
I think e.g. `test_name(32)` is a nice format. Also we start having a
standard, so we can change it everywhere if we agree on something else.
With this one I can quickly ctrl-f for `[32]` in most of tests in
`testdata`, and I find it pretty convenient for debugging.