Support Client certificate authentication for opensearch/elastic search  (#1406)
In addition to username and password, es/os support client-auth based
auth. See
[here](https://docs.opensearch.org/docs/latest/security/authentication-backends/client-auth/).
This pr wants to support that.

This change introduces client certificate authentication for Quesm when
connecting to Elasticsearch. It does not alter how users authenticate to
Quesm.

(this pr is to implement request of this
https://github.com/QuesmaOrg/quesma/issues/1394)
<!-- A note on testing your PR -->
<!-- Basic unit test run is executed against each commit in the PR.
If you want to run a full integration test suite, you can trigger it by
commenting
     with '/run-integration-tests' or '/run-it' -->

---------

Co-authored-by: przemyslaw <przemyslaw.hejman@quesma.com>