[done, needs some testing] Revert `Text/keyword` change? (#1386)
Before this PR all sample dashboards load in Kibana 8.11, but one single
panel doesn't in newest Kibana 8.17.
After this PR they all load in both of them.

https://github.com/QuesmaOrg/quesma/pull/1321

Draft for a bit, let me inspect a) PR above b) the customer case which
was why I also introduced the reverse change a few months ago (we used
to have exactly the same version in code as after this PR months ago)

---------

Co-authored-by: Przemek Delewski <pdelewski@quesma.com>