Rewrite index names on ingest (#1448)
This PR adds options to rewrite index names on ingest. It's beneficial
if we want all `daily` indices into single index.

It is configured as a list of rules. These rules are applied in order
(sorted names).

- `from` contains a matching regexp 
- `to` part contains a match group name


Rewriting is done at a very early stage of processing. 

Here is a example config:
```
  - name: my-ingest-processor
    type: quesma-v1-processor-ingest
    config:
      indexNameRewriteRules:
        0:
          from: (.*?)(-\d{4}\.\d{2}\.\d{2})$
          to: "$1"
        1:
          from: (.*?)(-\d{4}\.\d{2})$
          to: "$1"
        3:
          from: (.*?)(.\d{4}-\d{2})$
          to: "$1"
        4: 
          from: (.*?)(.\d{4}-\d{2}-\d{2})$
          to: "$1"  
```

It will rewrite `logs-2023.01` into `logs`.