Fix Slack notification not working for PRs from forks
I (and potential outside contributors) don't commit branches to the
QuesmaOrg/quesma repo, but to my personal fork. For such a workflow,
the issues-and-prs.yml job (Slack notifications for new PRs/issues)
didn't work, failing with error:

> Missing input! A token must be provided to use the method decided.

The problem is that the Slack token GitHub secret isn't available
for forks - due to security reasons.

However, by changing the type of job from `pull_request` to
`pull_request_target`, it changes how GitHub will run the job - it
will run it in context of "parent" repo (QuesmaOrg/quesma), meaning it
has access to the Slack token.