Add graceful configuration error (#1464)
Currently, Quesma is in a restart loop on configuration issues, which is
a terrible developer experience.

Instead of being able to read the error message, it keeps restarting.

My proposal instead is to return exit 0 and not restart. This allows
reading messages as well as gives a much faster error in `docker compose
up`.